archive/iso_audit_agent/requirements_iso.txt
streamlit
PyPDF2
python-docx
python-dotenv
google-generativeai
openai
pandas
plotly
